pyproject.toml
[project]
name = "lake"
version = "0.1.0"
description = "Data Lake initiator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bokeh>=3.8.0",
    "boto3==1.38.32",
    "briefer>=0.0.14",
    "clickhouse-driver>=0.2.9",
    "colorlog>=6.9.0",
    "confluent-kafka>=2.11.1",
    "duckdb>=1.4.1",
    "ducklake>=0.1.1",
    "ipykernel>=6.17.0",
    "bokeh==3.8.0",
    "ipywidgets>=8.1.7",
    "ipywidgets-bokeh==1.6.0",
    "setuptools==80.9.0",
    "kafka-python>=2.2.15",
    "matplotlib>=3.10.7",
    "numpy==2.2.6",
    "pandas==2.2.3",
    "panel>=1.8.2",
    "psycopg2-binary==2.9.10",
    "pyarrow==20.0.0",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "pytest==8.3.5",
    "python-dotenv==1.1.0",
    "pyyaml>=6.0.2",
    "apache-airflow[celery]==3.1.2",
    "apache-airflow-providers-google==18.0.0",
    "apache-airflow-providers-fab==3.0.0",
    "apache-airflow-providers-sftp==5.4.0",
    "celery==5.5.3",
    "asyncpg==0.30.0",
    "pendulum==3.1.0",
    "polars==1.34.0",
    "sqlalchemy==1.4.50",
    "duckdb-extension-postgres-scanner>=1.4.1"
]



[project.scripts]
lake = "lake.cmd:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
